Don't use Watch Connectivity Framework
This commit will get rid of the use of WC Framework for good. Instead of
doing API calls and location on iOS and passing the results via WC
messages to WatchOS, each target (iOS and WatchOS) does their own API
and location work for themselves.

Based on initial tests on the simulation, this is MUCH slower than the
WC message implementation, but makes the implementation MUCH simpler.
Need to test the performance some more before possibly merging to
master.