Get departures within next 6h. Inform user if no departures.
The API can return departures within a 6h timeframe at the longest. The
app will now search for departures within that 6h timeframe.

A more advanced solution is to search within the next 6h timeframe, if
no departures are found within the previous one. Will need to implement
this in the future.